Mudanças RBAC, novas permissões, Necessário refazer as migrações para aceder a Catálogo

+++ saramago/backend/controllers/CatController.php
@@ -76,22 +76,25 @@ class CatController extends Controller
      * @return mixed
      */
     public function actionIndex()
-    {
-        $this->layout="minor";
+    {   
+        if ((Yii::$app->user->can('acessoCatalogo'))) {
+            $this->layout="minor";
 
-        $obrasModel = Obra::find()->all();
-        $exemplarModels = Exemplar::find()->all();
-        $cduAll = ArrayHelper::map(Cdu::find()->all(),'id','designacao','codCdu',['enctype' => 'multipart/form-data']);
-        $tiposExemplarAll = ArrayHelper::map(Tipoexemplar::find()->all(),'id','designacao','tipo',['enctype' => 'multipart/form-data']);
-        $colecaoAll = ArrayHelper::map(Colecao::find()->all(), 'id', 'tituloColecao', ['enctype' => 'multipart/form-data']);
+            $obrasModel = Obra::find()->all();
+            $exemplarModels = Exemplar::find()->all();
+            $cduAll = ArrayHelper::map(Cdu::find()->all(),'id','designacao','codCdu',['enctype' => 'multipart/form-data']);
+            $tiposExemplarAll = ArrayHelper::map(Tipoexemplar::find()->all(),'id','designacao','tipo',['enctype' => 'multipart/form-data']);
+            $colecaoAll = ArrayHelper::map(Colecao::find()->all(), 'id', 'tituloColecao', ['enctype' => 'multipart/form-data']);
 
-        $searchModel = new ObraSearch();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+            $searchModel = new ObraSearch();
+            $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
-        return $this->render('index',
-            ['searchModel' => $searchModel, 'dataProvider' => $dataProvider,
-                'obrasModel' => $obrasModel, 'cduAll'=>$cduAll,
-                'tiposExemplarAll'=>$tiposExemplarAll, 'colecaoAll' => $colecaoAll]);
+            return $this->render('index',
+                ['searchModel' => $searchModel, 'dataProvider' => $dataProvider,
+                    'obrasModel' => $obrasModel, 'cduAll'=>$cduAll,
+                    'tiposExemplarAll'=>$tiposExemplarAll, 'colecaoAll' => $colecaoAll]);
+        }
+        return $this->redirect(['../web']);
     }
 
     /**
@@ -102,7 +105,10 @@ class CatController extends Controller
      */
     public function actionView($id)
     {
-        return $this->renderAjax('view', ['model' => $this->findModel($id),]);
+        if ((Yii::$app->user->can('verObras'))) {
+            return $this->renderAjax('view', ['model' => $this->findModel($id),]);
+        }
+        return $this->redirect(['../web']);
     }
 
     /**
@@ -113,18 +119,21 @@ class CatController extends Controller
      */
     public function actionViewFull($id)
     {
-        $this->layout = 'minor';
+        if ((Yii::$app->user->can('verObras'))) {
+            $this->layout = 'minor';
 
-        $bibliotecaAll = ArrayHelper::map(Biblioteca::find()->all(),'id','nome',['enctype' => 'multipart/form-data']);
-        //for ->actionExemplarView
-        $totalExemplaresDaObra = Exemplar::find()->where('Obra_id='.$id)->count();
-        $exemplarModels = Exemplar::find()->all();
-        $searchModel = new ExemplarSearch();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+            $bibliotecaAll = ArrayHelper::map(Biblioteca::find()->all(),'id','nome',['enctype' => 'multipart/form-data']);
+            //for ->actionExemplarView
+            $totalExemplaresDaObra = Exemplar::find()->where('Obra_id='.$id)->count();
+            $exemplarModels = Exemplar::find()->all();
+            $searchModel = new ExemplarSearch();
+            $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
 
-        return $this->render('view-full', ['model' => $this->findModel($id), 'totalExemplaresDaObra' => $totalExemplaresDaObra,
-            'searchModel' => $searchModel, 'dataProvider' => $dataProvider, 'exemplarModels' => $exemplarModels, 'bibliotecaAll' => $bibliotecaAll]);
+            return $this->render('view-full', ['model' => $this->findModel($id), 'totalExemplaresDaObra' => $totalExemplaresDaObra,
+                'searchModel' => $searchModel, 'dataProvider' => $dataProvider, 'exemplarModels' => $exemplarModels, 'bibliotecaAll' => $bibliotecaAll]);
+        }
+        return $this->redirect(['../web']);
     }
 
     /**
@@ -134,19 +143,22 @@ class CatController extends Controller
      */
     public function actionCreate()
     {
-        $model = new ObraForm();
+        if ((Yii::$app->user->can('inserirObras'))) {
+            $model = new ObraForm();
 
-        $cduAll = ArrayHelper::map(Cdu::find()->all(),'id','designacao','codCdu',['enctype' => 'multipart/form-data']);
-        $colecaoAll = ArrayHelper::map(Colecao::find()->all(),'id','tituloColecao',['enctype' => 'multipart/form-data']);
-        if(Yii::$app->request->isPost) {
-            $model->imageFile = UploadedFile::getInstance($model, 'imageFile');
-            if ($model->createObra()) {
-                Yii::$app->session->setFlash('success', "<strong>Informação:</strong> O " . $model->titulo . " foi adicionado.");
-                return $this->redirect(['view-full', 'id' => $model->id]);
+            $cduAll = ArrayHelper::map(Cdu::find()->all(),'id','designacao','codCdu',['enctype' => 'multipart/form-data']);
+            $colecaoAll = ArrayHelper::map(Colecao::find()->all(),'id','tituloColecao',['enctype' => 'multipart/form-data']);
+            if(Yii::$app->request->isPost) {
+                $model->imageFile = UploadedFile::getInstance($model, 'imageFile');
+                if ($model->createObra()) {
+                    Yii::$app->session->setFlash('success', "<strong>Informação:</strong> O " . $model->titulo . " foi adicionado.");
+                    return $this->redirect(['view-full', 'id' => $model->id]);
+                }
             }
-        }
 
-        return $this->renderAjax('create', ['model' => $model, 'cduAll'=> $cduAll, 'colecaoAll' => $colecaoAll]);
+            return $this->renderAjax('create', ['model' => $model, 'cduAll'=> $cduAll, 'colecaoAll' => $colecaoAll]);
+        }
+        return $this->redirect(['../web']);
     }
 
     /**
@@ -158,13 +170,16 @@ class CatController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if ((Yii::$app->user->can('editarObras'))) {
+            $model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
+
+            return $this->renderAjax('update', ['model' => $model]);
         }
-
-        return $this->renderAjax('update', ['model' => $model]);
+        return $this->redirect(['../web']);
     }
 
     /**
@@ -175,10 +190,13 @@ class CatController extends Controller
      * @throws NotFoundHttpException if the model cannot be found
      */
     public function actionDelete($id)
-    {
-        $this->findModel($id)->delete();
+    {  
+        if ((Yii::$app->user->can('editarObras'))) {
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        }
+        return $this->redirect(['../web']);
     }
 
     /**
@@ -206,11 +224,14 @@ class CatController extends Controller
      */
     public function actionExemplarIndex()
     {
-        $searchModel = new ExemplarSearch();
-        
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+        if ((Yii::$app->user->can('verExemplares'))) {
+            $searchModel = new ExemplarSearch();
+            
+            $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
-        return $this->render('exemplar/index', ['searchModel' => $searchModel, 'dataProvider' => $dataProvider]);
+            return $this->render('exemplar/index', ['searchModel' => $searchModel, 'dataProvider' => $dataProvider]);
+        }
+        return $this->redirect(['../web']);
     }
 
     /**
@@ -221,7 +242,10 @@ class CatController extends Controller
      */
     public function actionExemplarView($id)
     {
-        return $this->renderAjax('exemplar/view', ['model' => $this->findModelExemplar($id),]);
+        if ((Yii::$app->user->can('verExemplares'))) {
+            return $this->renderAjax('exemplar/view', ['model' => $this->findModelExemplar($id),]);
+        }
+        return $this->redirect(['../web']);
     }
 
     /**
@@ -231,17 +255,20 @@ class CatController extends Controller
      */
     public function actionExemplarCreate()
     {
-        $model = new ExemplarForm();
+        if ((Yii::$app->user->can('inserirExemplares'))) {
+            $model = new ExemplarForm();
 
-        $estatutoexemplarAll = ArrayHelper::map(Estatutoexemplar::find()->all(),'id','estatuto',['enctype' => 'multipart/form-data']);
-        $tipoexemplarAll = ArrayHelper::map(Tipoexemplar::find()->all(),'id','designacao','tipo',['enctype' => 'multipart/form-data']);
-        $bibliotecaAll = ArrayHelper::map(Biblioteca::find()->all(),'id','nome',['enctype' => 'multipart/form-data']);
+            $estatutoexemplarAll = ArrayHelper::map(Estatutoexemplar::find()->all(),'id','estatuto',['enctype' => 'multipart/form-data']);
+            $tipoexemplarAll = ArrayHelper::map(Tipoexemplar::find()->all(),'id','designacao','tipo',['enctype' => 'multipart/form-data']);
+            $bibliotecaAll = ArrayHelper::map(Biblioteca::find()->all(),'id','nome',['enctype' => 'multipart/form-data']);
 
-        if ($model->load(Yii::$app->request->post()) && $model->createExemplar()) {
-            Yii::$app->session->setFlash('success', "<strong>Informação:</strong> Um novo exemplar foi adicionado.");
-            return $this->redirect(['view-full', 'id' => $model->obra->id]);
+            if ($model->load(Yii::$app->request->post()) && $model->createExemplar()) {
+                Yii::$app->session->setFlash('success', "<strong>Informação:</strong> Um novo exemplar foi adicionado.");
+                return $this->redirect(['view-full', 'id' => $model->obra->id]);
+            }
+            return $this->renderAjax('exemplar/create', ['model' => $model, 'bibliotecaAll' => $bibliotecaAll, 'tipoexemplarAll' => $tipoexemplarAll, 'estatutoexemplarAll' => $estatutoexemplarAll]);
         }
-        return $this->renderAjax('exemplar/create', ['model' => $model, 'bibliotecaAll' => $bibliotecaAll, 'tipoexemplarAll' => $tipoexemplarAll, 'estatutoexemplarAll' => $estatutoexemplarAll]);
+        return $this->redirect(['../web']);
     }
 
     /**
@@ -253,17 +280,20 @@ class CatController extends Controller
      */
     public function actionExemplarUpdate($id)
     {
-        $model = $this->findModelExemplar($id);
+        if ((Yii::$app->user->can('editarExemplares'))) {
+            $model = $this->findModelExemplar($id);
 
-        $estatutoexemplarAll = ArrayHelper::map(Estatutoexemplar::find()->all(),'id','estatuto',['enctype' => 'multipart/form-data']);
-        $tipoexemplarAll = ArrayHelper::map(Tipoexemplar::find()->all(),'id','designacao','tipo',['enctype' => 'multipart/form-data']);
-        $bibliotecaAll = ArrayHelper::map(Biblioteca::find()->all(),'id','nome',['enctype' => 'multipart/form-data']);
+            $estatutoexemplarAll = ArrayHelper::map(Estatutoexemplar::find()->all(),'id','estatuto',['enctype' => 'multipart/form-data']);
+            $tipoexemplarAll = ArrayHelper::map(Tipoexemplar::find()->all(),'id','designacao','tipo',['enctype' => 'multipart/form-data']);
+            $bibliotecaAll = ArrayHelper::map(Biblioteca::find()->all(),'id','nome',['enctype' => 'multipart/form-data']);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view-full', 'id' => $model->obra->id]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view-full', 'id' => $model->obra->id]);
+            }
+
+            return $this->renderAjax('exemplar/update', ['model' => $model, 'bibliotecaAll' => $bibliotecaAll, 'tipoexemplarAll' => $tipoexemplarAll, 'estatutoexemplarAll' => $estatutoexemplarAll]);
         }
-
-        return $this->renderAjax('exemplar/update', ['model' => $model, 'bibliotecaAll' => $bibliotecaAll, 'tipoexemplarAll' => $tipoexemplarAll, 'estatutoexemplarAll' => $estatutoexemplarAll]);
+        return $this->redirect(['../web']);
     }
 
     /**
@@ -275,15 +305,18 @@ class CatController extends Controller
      */
     public function actionExemplarDelete($id)
     {
-        $model = Exemplar::findOne($id);
-        if ($model->estado == 'nd' || $model->estado == 'estante') {
-            $this->findModelExemplar($id)->delete();
-        }
-        else {
-            Yii::$app->session->setFlash('error', "Erro ao remover o exemplar: Só é possível remover exemplares no estado 'Não disponível' ou 'Na estante' ");
-        }
+        if ((Yii::$app->user->can('eliminarExemplares'))) {
+            $model = Exemplar::findOne($id);
+            if ($model->estado == 'nd' || $model->estado == 'estante') {
+                $this->findModelExemplar($id)->delete();
+            }
+            else {
+                Yii::$app->session->setFlash('error', "Erro ao remover o exemplar: Só é possível remover exemplares no estado 'Não disponível' ou 'Na estante' ");
+            }
 
-        return $this->redirect(['view-full', 'id' => $model->obra->id]);
+            return $this->redirect(['view-full', 'id' => $model->obra->id]);
+        }
+        return $this->redirect(['../web']);
     }
 
     /**
